feat(workspace): split into trash-rs and trash-rs-cli, prepare crates.io metadata
Convert the repo to a Cargo workspace with separate library and CLI crates, update CI/pre-commit to workspace commands, and add crate-specific README/Cargo metadata for publishing.

+++ crates/trash-rs/src/trash/macos.rs
@@ -18,7 +18,7 @@ pub(crate) fn move_files_to_trash(paths: &[PathBuf]) -> Result<()> {
             return Err(TrashError::TrashOperation {
                 path: path.clone(),
                 domain: error.domain().to_string(),
-                code: error.code() as isize,
+                code: error.code(),
                 message: error.localizedDescription().to_string(),
             });
         }
